s8_deployment/exercise_files/requirements.txt
numpy==1.21.2
protobuf==3.19.3
requests==2.26.0
scikit_learn==1.0.2
torch==1.10.1
torchvision==0.11.2
google-cloud-storage